refactor(dynclassification): remove StandardDocParser alias, fix LLM dtype, add MLX conversion script and ML deps
- Remove StandardDocParser backward-compat alias and update all imports/docs.
- Fix Qwen2VL model loading: use torch_dtype instead of dtype.
- Add scripts/convert_models_to_mlx.py for Apple Silicon Metal inference.
- Enhance standard_profile_generator negation filtering and add regression tests.
- Install CPU torch + ML extras; GPU wheel install blocked by slow network timeouts.

+++ privacy_local_agent/dynclassification/llm_engines.py
@@ -203,7 +203,7 @@ class Qwen2VLClassifier(LlmClassifier):
                 # 从本地目录加载预训练模型权重
                 self._model = Qwen2VLForConditionalGeneration.from_pretrained(
                     self.model_path,
-                    dtype=dtype,  # 模型精度
+                    torch_dtype=dtype,  # 模型精度
                     # CUDA/MPS 使用自动设备映射（多层分配到不同设备），CPU 不需要
                     device_map="auto" if device in ("cuda", "mps") else None,
                 )
